nodes/Node_95_WebRTC_Receiver/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pillow==10.1.0
aiortc==1.6.0
av==11.0.0
